old/code/siamese_torch/requirements.txt
absl-py==2.0.0
aggdraw==1.3.18.post0
aiohttp==3.9.1
aiosignal==1.3.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorcet==3.1.0
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
datasets==2.16.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
dm-tree==0.1.8
docker-pycreds==0.4.0
exceptiongroup==1.0.4
executing==0.8.3
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.10.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.26.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.60.0
h5py==3.10.0
huggingface-hub==0.20.2
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipydatawidgets==4.3.2
ipykernel==6.28.0
ipympl==0.9.3
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==7.8.1
itk-core==5.3.0
itk-filtering==5.3.0
itk-meshtopolydata==0.10.0
itk-numerics==5.3.0
itkwidgets==0.32.6
jedi==0.18.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
jupyterlab-widgets==1.1.7
keras==3.0.2
kiwisolver==1.4.5
libclang==16.0.6
llvmlite==0.42.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
namex==0.0.7
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.5.8
networkx==3.2.1
nibabel==5.2.1
notebook==6.4.13
numba==0.59.0
numpy==1.23.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.16.0
onnxscript==0.1.0.dev20240401
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==4.1.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
Pygments==2.15.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytorchsummary==1.3.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==1.45.0
setproctitle==1.3.3
setuptools==68.2.2
SimpleITK==2.3.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.2.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.2.0
tinycss2==1.2.1
torch==2.1.2
torchaudio==2.1.2
torchview==0.2.6
torchvision==0.16.2
torchviz==0.0.2
tornado==6.4
tqdm==4.66.1
traitlets==5.6.0
traittypes==0.2.1
triton==2.1.0
typing_extensions==4.7.1
tzdata==2023.4
urllib3==2.1.0
visualkeras==0.0.2
wandb==0.17.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.41.2
widgetsnbextension==3.6.6
wrapt==1.14.1
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
